Move runtime parts into runtime mod
